terminando funcionalidades
