+++ public/index.html
@@ -28,9 +28,9 @@
     <script id="react-dotenv" src="./env.js"></script>
   </head>
 
-  <body>
+  <body id="root">
     <noscript>You need to enable JavaScript to run this app.</noscript>
-    <div id="root"></div>
+    <div></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
