+++ src/components/button/buttonComponent.js
@@ -1,9 +1,10 @@
 import './style.css'
-export const Button = ({ text = '', onClick }) => {
+export const Button = ({ text, onClick, icon }) => {
   return (
     <>
       <div>
-        <button onClick={onClick}>{text}</button>
+        <button onClick={onClick}>
+          {text}</button>
       </div>
     </>
   )
