+++ src/components/forms/movieForm.js
@@ -1,0 +1,38 @@
+
+import { Button } from '../button/buttonComponent'
+import { Form } from './form'
+import { Input } from '../input/inputComponent'
+// import { client } from '../../../service/client'
+
+export const MovieForm = ({ textButton, onClick, onChange }) => {
+  return (
+    <>
+      <Form>
+        <div>
+          <Input name='title' type='text' label='Title' onChange={onChange} />
+          <Input name='subtitle' type='text' label='Subtitle' onChange={onChange} />
+        </div>
+        <div>
+          <Input name='resume' type='text' label='Resume' onChange={onChange} />
+        </div>
+        <div>
+          <Input name='gender' type='text' label='Gender' onChange={onChange} />
+          <Input name='classification' type='text' label='Classification' onChange={onChange} />
+        </div>
+        <Input name='releaseDate' type='text' label='Release Date' onChange={onChange} />
+        <Input name='image' type='text' label='Image' onChange={onChange} />
+        <div>
+          <Input name='director' type='text' label='Director' onChange={onChange} />
+          <Input name='writer' type='text' label='writer' onChange={onChange} />
+        </div>
+        <Input name='studio' type='text' label='Studio' onChange={onChange} />
+        <Input name='note' type='text' label='Stars' onChange={onChange} />
+        <div>
+          <Input name='actors' type='text' label='actors' onChange={onChange} />
+          <Input name='awards' type='text' label='awards' onChange={onChange} />
+        </div>
+        <Button text={textButton} onClick={onClick} />
+      </Form>
+    </>
+  )
+}
