+++ src/components/modal/modal.js
@@ -2,14 +2,14 @@ import { Button } from '../button/buttonComponent'
 import { useState } from 'react'
 import './modal.css'
 
-export const Modal = ({ children, title }) => {
+export const Modal = ({ children, title, buttonStyle }) => {
   const [toggle, setToggle] = useState(false)
 
   const handleClick = () => setToggle(!toggle)
 
   return (
     <>
-      <Button text="Abrir Modal" onClick={handleClick} />
+      <Button text="Abrir Modal" onClick={handleClick} buttonStyle={buttonStyle} />
 
       {toggle && (
         <div className='modal-background'>
