--- src/pages/home/home.css
@@ -1,6 +1,0 @@
-.movieList {
-  margin: 50%;
-  padding: 50%;
-  justify-content: center;
-  align-items: center;
-}
